added random names to db and across UI
