+++ tic-tac-toe/drizzle/0001_odd_lord_tyger.sql
@@ -1,0 +1,1 @@
+ALTER TABLE "games" ADD COLUMN "name" varchar(255) NOT NULL;
