+++ tic-tac-toe/drizzle/meta/_journal.json
@@ -8,6 +8,13 @@
       "when": 1749076619863,
       "tag": "0000_fancy_chamber",
       "breakpoints": true
+    },
+    {
+      "idx": 1,
+      "version": "7",
+      "when": 1749160346395,
+      "tag": "0001_odd_lord_tyger",
+      "breakpoints": true
     }
   ]
 }
