+++ tic-tac-toe/src/Lobby.tsx
@@ -3,37 +3,6 @@ import type { Game, Player } from "./game/game"
 import { useState } from "react"
 import { ClientTicTacToe } from "./api"
 
-function generateGameName() {
-  const nameList = [
-    'Time','Past','Future','Dev',
-    'Fly','Flying','Soar','Soaring','Power','Falling',
-    'Fall','Jump','Cliff','Mountain','Rend','Red','Blue',
-    'Green','Yellow','Gold','Demon','Demonic','Panda','Cat',
-    'Kitty','Kitten','Zero','Memory','Trooper','XX','Bandit',
-    'Fear','Light','Glow','Tread','Deep','Deeper','Deepest',
-    'Mine','Your','Worst','Enemy','Hostile','Force','Video',
-    'Game','Donkey','Mule','Colt','Cult','Cultist','Magnum',
-    'Gun','Assault','Recon','Trap','Trapper','Redeem','Code',
-    'Script','Writer','Near','Close','Open','Cube','Circle',
-    'Geo','Genome','Germ','Spaz','Shot','Echo','Beta','Alpha',
-    'Gamma','Omega','Seal','Squid','Money','Cash','Lord','King',
-    'Duke','Rest','Fire','Flame','Morrow','Break','Breaker','Numb',
-    'Ice','Cold','Rotten','Sick','Sickly','Janitor','Camel','Rooster',
-    'Sand','Desert','Dessert','Hurdle','Racer','Eraser','Erase','Big',
-    'Small','Short','Tall','Sith','Bounty','Hunter','Cracked','Broken',
-    'Sad','Happy','Joy','Joyful','Crimson','Destiny','Deceit','Lies',
-    'Lie','Honest','Destined','Bloxxer','Hawk','Eagle','Hawker','Walker',
-    'Zombie','Sarge','Capt','Captain','Punch','One','Two','Uno','Slice',
-    'Slash','Melt','Melted','Melting','Fell','Wolf','Hound',
-    'Legacy','Sharp','Dead','Mew','Chuckle','Bubba','Bubble','Sandwich',
-    'Smasher','Extreme','Multi','Universe','Ultimate','Death','Ready','Monkey',
-    'Elevator','Wrench','Grease','Head','Theme','Grand','Cool','Kid',
-    'Boy','Girl','Vortex','Paradox'
-  ]
-
-  return `${nameList[Math.floor( Math.random() * nameList.length)] + " " + nameList[Math.floor( Math.random() * nameList.length)] }`
-}
-
 export function Lobby() {
   const { ongoingGames: loadedGames } = useLoaderData<{ ongoingGames: Game[] }>()
   const [games, setGames] = useState<Game[]>(loadedGames)
@@ -58,10 +27,9 @@ export function Lobby() {
       <h3 className="text-lg font-medium">JOIN A GAME</h3>
       <ul className="p-0 m-0 w-fit mx-auto">
         {games.map(game => (
-          <li className="list-decimal text-blue-500 hover:underline"><Link to={`/game/${game.id}`}>{generateGameName()}</Link></li>
+          <li key={game.id} className="list-decimal text-blue-500 hover:underline"><Link to={`/game/${game.id}`}>{game.name}</Link></li>
         ))}
       </ul>
-
     </div>
   )
 }
